more unit tests

+++ server/controllers/adsController.js
@@ -183,6 +183,13 @@ class AdsController {
         });
         return;
       }
+      if (car.rows[0].owner !== req.user.id && req.user.email !== 'createadmin@yahoo.com') {
+        res.status(403).json({
+          status: 403,
+          error: 'Sorry, you can not delete this car',
+        });
+        return;
+      }
       const deleteCar = 'DELETE FROM cars WHERE id = $1';
       await db.query(deleteCar, [value]);
